feat(examples): last changes to be video demo ready
